+++ spec/contract/documentation_spec.rb
@@ -1,7 +1,7 @@
 require "rbconfig"
 
 RSpec.describe "public documentation" do
-  it "documents public classes and required version 0.5 topics" do
+  it "documents public classes and required version 0.6 topics" do
     script = File.expand_path("../../script/verify_docs", __dir__)
 
     expect(system(RbConfig.ruby, script)).to eq(true)